route helper tests
